Remove arm64v8
On many builds, we get:

> `pypy3: error while loading shared libraries: libffi.so.6: cannot open shared object file: No such file or directory`

Which we can't resolve on Bullseye+ (same as s390x).

On others, we get:

> `pypy: error while loading shared libraries: libtinfow.so.6: cannot open shared object file: No such file or directory`

Which appears to be a CentOS-specific library name and I couldn't convince `pypy` to be satisfied by any symlinks to either `libtinfo` or `libtic` (which were the suggestions I found on other projects for fixing similar linking issues).